Another Mathematica NB upload.
I'm uploading a related notebook. This one is shorter (except for some data input, which is not visible on the images). It will be a nice look at some other real data. I have the `*.nb` files, a PDF of the evaluated Notebook, and images of the unevaluated and evaluated notebooks.